requirements.txt
asgiref==3.8.1
Django==5.2.3
django-environ==0.12.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2020.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.4.1
tzdata==2025.2
